Allow user to configure "owning" references (#9)
## Issue Description

In many scenarios, the same object can be referenced from several other objects (such as caches or parent pointers) in addition to the object that is the primary "owner". This can make for cluttered backtraces, and - worse - make many objects float all the way up to the top of the dominator tree.

## Change Description

With this option, the user can provide a configuration file that marks certain object fields as "owning" references. If an object is found to be referenced by both owning and non-owning references, the non-owning references are discarded from backtraces (and, hence, are ignored for construction of the dominator tree).

## Test Plan

Compare differences in the treemap with and without a configuration file.