Support owning references via collections (#13)
## Issue Description

There are many cases where a collection type is used to hold owning references. Previously, it was possible only to mark all collections of a certain type as owning, when, often, only collections of that type in a specific field should be considered owning.

## Change Description

Added a new kind of field rule to represent deep matching, as described in `README.md`.