Leak detection tool: logic to detect "unowned" objects (#16)
## Issue Description

In the service-oriented architecture of the product whose memory usage we're analyzing, we have a lot of types that are logical singletons in that there should be only one instance of them at any given point in time (though that instance can be recreated over time, so they are not true singletons).

We have seen scenarios where on recreation, the previous instance did not unsubscribe itself properly from notifications, and stayed around when it should have become eligible for garbage collection.

It is interesting to us to be able to automatically analyze a dump to report when this happened.

## Change Description

It turns out that we can use our existing concept of reference classifiers to build a leak detection tool.

The idea is as follows:
* Craft a reference classifier that identifies the "official" singleton instance at any given point in time.
* Run a backtrace analysis using this reference classifier.
* Enumerate all types for which we have found "owned" instances, according to the reference classifier.
* Then, enumerate all instances of these types that are not "owned", according to the reference classifier.

The command this logic is embodied in is `listobj 'unowned`.

## Bonus Changes

To help with diagnosing the reported objects, added some minor additional features:
* Added a flag `'includederived` to both `dumptype` and `listobjs 'type`. With this, we can do things like produce stats on boxing (e.g., run `listobjs 'includederived 'type` on the type index of `System.ValueType`).
* Added options to dump all owned/unowned instances of a given type. This can be useful when we want to make sure that all instances of a type are properly parented in the dominator tree, but we need several reference classifier rules to achieve that.
* Added an `'exec` option to `listobjs`. With this, we can quickly do things like dumping objects (`listobjs 'exec "dumpobj 'memory"`) or showing backtraces for the selected objects (`lo 'unowned 'exec "bt 'depth 3"`).
* Added `options 'noreferenceclassifier` to reset the reference classifier on a given context.
* `heapdom 'relativeto` now produces better annotations about the differences found, for more helpful visualization in the treemap viewer.

## Test Plan

Ran the leak detection tool against a memory snapshot. Confirmed several of the reported instances as indeed having bugs. This uncovered other variants of leaks than just event subscriptions, so the command does provide some good general-purpose value.