Add 'weakdelegates option (#20)
## Issue Description

Delegates can hold on to references (and, possibly, large object graphs)
via their closures, and delegates are often used to connect otherwise
independent subsystems for notification delivery. This means that
references to other objects from delegate closures usually do not
contribute to the "ownership" relationship between objects and are more
incidental.

## Change Description

With the `'weakdelegates` option, references from delegate closures are
treated as "weak" -- if objects are reachable both from other object
fields and from delegate closures, we will ignore the delegate closure
reference. If an object is reachable *only* from the delegate closure,
it is a likely leak candidate, which will be shown with a `..` prefix in
backtraces.

Also, renamed the `fusegchandles` option to `fuseroots`, to better
reflect what it does.

## Test Plan

* Confirmed that backtraces taken during leak analysis are less
cluttered and still useful.
* Confirmed that the dominator tree has fewer nodes that get pulled all
the way up to the root node.