Chasing memory leaks with reference classifiers and "lifelines" (#26)
## Issue Description

Reference classifiers, together with the `listobj 'unowned` command,
have proven to be valuable tools for identifying objects that should
(but haven't been) eligible for garbage collection. However, in some
cases it could still be cumbersome to spot the reference that needed to
be broken to make the object unreachable, in particular in deep, highly
interlinked object graphs.

## Change Description

This change provides a new command, `backtrace 'lifelines`, that
produces a compressed view of representative (and, computed as a cheap
approximation, short) reference chains to the given (leaked) object that
need to be broken. The view is compressed as follows:
* Stop at objects that are "strongly owned", according to the reference
classifier.
* Keep a single reference chain per distinct root from which the target
object can be reached. While we don't enumerate, internally, all
reference chains from the root to the target, we display only the
shortest one among those we did enumerate.
* Conserve indentation in the presence of deep single-reference chains,
and prefix the line with a `\` continuation character instead.

Added a section to the `README.md` with recommendations for a process to
chase down memory leaks using reference classifiers and lifelines.

Removed some options from the `backtrace` command that outlived their
usefulness at this point, where better tools exist in the command suite.