Easy identification of destroyed/disposed objects (#31)
## Issue Description

When identifying leaks, it can be helpful to quickly see (e.g., in
backtraces) when an object has been disposed.

## Change Description

Introducing "tag" rules, that allow objects to be given a specific tag
if a specified field has value 0 (or is non-zero). The tag is displayed
as part of object listings and backtraces.