Add support for dynamic selectors (#33)
## Issue Description

For simplicity and performance, selectors (as used in "owns" rules) were
bound to type indices and field numbers statically. In a few cases, this
is not enough to express the right ownership relationships - e.g., when
a field is declared as an `IList`, the actual list items are not
available statically as a field but require dynamic lookup.

## Change Description

Use static lookup for as much of the selector prefix as possible, then
switch to dynamic lookup for the remainder. This makes this the
exceptional case, and is pay-for-play in terms of performance.