Allow saving of reference classifiers
Summary: In the future, we'll have other data sources for reference classifiers than text files. Once we wire those up, it will be useful to be able to serialize reference classifiers to text files, which can be used decoupled from the other data sources. This change introduces `rc 'save` to allow saving.

Differential Revision: D48607065

fbshipit-source-id: 8be7781d5974ffc0d6197938a5cceb8cc5adbdd2