Define attributes to allow specifying reference classifiers in-source
Summary:
So far, we only supported writing reference classification rules in text files separate from the product under analysis. This made it possible for the product code to be churned, and the reference classifiers to get out of sync.

This change introduces custom attributes that can be deployed into product code. In a followup change, we'll teach MemorySnapshotAnalyzer to extract rules from managed DLLs.

Reference classifier text files are going to continue to be relevant for DLLs not built from source (including the .NET framework).

Differential Revision: D48607066

fbshipit-source-id: 928d70fc4f48d0fa886e5efa3b3cb02e60c03ecc