Add warnings when selectors find unmapped memory, plus "dumpobj 'selector"
Summary:
We have been observing that Unity memory snapshots can sometimes be missing some heap segments (which results in invalid pointers reported during tracing, and with the `dumpinvalidrefs` command). So far, this has been pretty benign, but recently we ran into a case where the backing store of a collection of critical importance for memory attribution was missing. As we set out to collect snapshots and do analysis in automation, we want to make sure that we can easily recognize bad signals (snapshots that are missing critical data) so we can quickly discard them.

To this aim, this change logs warnings when selectors (of "owns" or "tag" rules) find invalid references. These warnings are properly plumbed through to get into our logger objects, which we can dump and inspect separately in automation.

Also, this change adds a new option to `dumpobj`, called `'selector`. With this option, we can evaluate selectors interactively from the command line to diagnose problems more easily. Also, this allows one to manually test selectors when writing reference classifier rules by hand.

Reviewed By: jasonxzfb

Differential Revision: D48995039

fbshipit-source-id: 3650a16d44bed9bcd7b62dea2b322d7e8bf81c65