Add unit tests for TypeSystem.BindSelectors, and start harnessing SegmentedHeap.InterpretSelectors
Summary:
Now that `BindSelectors` is on `TypeSystem` (an outcome of the previous change), we can easily unit test the method. Also, this brings us back to 100% block coverage on `TypeSystem`.

Also, adding the required mocks and test objects that will enable us to test `SegmentedHeap.InterpretSelector`, but the tests have yet to be added.

Reviewed By: jasonxzfb

Differential Revision: D49073642

fbshipit-source-id: 59aa8631eb5b0be3d3e5813a51bfdfb560e7ff1c