Improve dumpinvalidrefs for diagnosing unmapped memory issues
Summary:
`dumpinvalidrefs` was just dumping objects with invalid references in the order they were encountered during tracing. Now, they are grouped by type index, repeated occurrences are elided unless `'verbose` is passed, and we also print the field name leading to the invalid reference. All of this will help diagnose the impact of missing heap segments in snapshots better.

While doing this I noticed that external references were reported as invalid references when not tracing a stitched heap; fixed this to remove the clutter from `dumpinvalidrefs`.

Differential Revision: D49118222

fbshipit-source-id: a145ff3b83f513ac65a50e553cec55f5c9e42675