Condense output when there are a lot of warnings
Summary:
As it turns out, now that we are reporting warnings when a selector leads into memory that is not part of the memory snapshot, we do regularly see some heavy hitters. This clutters the output quite a bit when using the tool interactively.

In this change, we are removing the straight-to-console warning logger, and extend the memory logger to be able to provide a summary of warnings. This will allow us to determine more quickly whether some critical piece of memory seems to be missing, making analysis results on that particular memory snapshot unusable.

Also, changed the format of the warning messages to have more pieces in common, to make scanning of full verbose logs more efficient as well.

Reviewed By: jasonxzfb

Differential Revision: D49106457

fbshipit-source-id: 43f31fa61f6f76736a565d8d4269de0eaa0edb50