Force-parent nodes whose backtraces don't lead to any roots into dominator tree
Summary:
When dealing with references of different weights, we can end up with objects that are reachable via a heap trace, but whose backtraces do not lead to *any* roots. For instance, an object can be reachable via a back-edge whose weight is higher than the references from a root, such as a weak GC handle.

Such objects would end up not being part of the dominator tree at all. Visualizing the heap as a treemap would, therefore, paint an incomplete picture of the state of the heap. This is particular irksome as leak scenarios can end up with some leaked objects becoming unreachable this way, and if we use the dominator tree for memory attribution, the leak wouldn't be reported.

To prevent this, with this change we are introducing (specifically in dominator tree computation) a node called "Unreachable" that is a direct child of the root under which we parent all nodes that would otherwise be missing from the dominator tree.

Differential Revision: D49161908

fbshipit-source-id: f421b5640ce615cbabdf2a52f2edda144f9e9f25