Option to allow bt 'lifelines to continue past strongly-owned objects
Summary:
When running with a large set of reference classifiers (and not a trimmed-down set for leak detection), it can still be useful to use the `bt 'lifelines` command. However, often that command will truncate lifelines to be shorter than is helpful because they will stop at any strongly-owned object. With this change, we are adding an option `'toroots` that provides the full lifelines, not stopping at strongly-owned objects.

Also includes some simple additions to reference classifiers for framework types.

Reviewed By: jasonxzfb

Differential Revision: D49334070

fbshipit-source-id: 8395168fd81786e45bcac50f4c0a7f4174622181