Add ignorebackref command to support diagnosing multiple memory leaks
Summary:
When a single memory leak causes a large number of unowned or destroyed objects to still be reachable, it can be cumbersome to confirm that there are no other leaks in a given memory snapshot. This change introduces a new command `ignorebackref` that can be used to ignore references to an object (from a specific parent object, or from any object) when computing backtraces.

When running a command such as `lo 'unowned 'exec "bt 'lifelines"` after breaking a reference in the lifelines, all leaked objects will still be listed, but (unless they are reachable through other lifelines) will have an empty lifeline. This makes it easy to spot leaks that aren't caused by the given reference, speeding up workflows when fixing multiple memory leaks in one go.

Differential Revision: D50019098

fbshipit-source-id: cad5abe0235a456a43933fd8883181fa2c4a53a5