Add flag to heapdom command to not launch the resulting HTML file
Summary: To make it easy to investigate baseline regressions in memory usage, we'll want automation to produce a heapdom treemap. When MemorySnapshotAnalyzer was just used interactively, by default we launched a browser to display the resulting HTML file. Now, for automation, we want a flag that prevents a browser from being launched.

Differential Revision: D54434330

fbshipit-source-id: 664189da48b463f058c31c1ee0ffe6b657b1f826