Support filtering heapdomstats by types, to support attribution of targeted sets of objects
Summary:
Now that we want to push higher attribution for specific features, it's helpful to be able to list unattributed objects that float to the top of the dominator tree, and that need more reference classifiers.

Note that this is not sufficient to ensure complete attribution, as matching objects could still be descendants of other top-level objects, or float to non-meaningful intermediate nodes.

Differential Revision: D54563606

fbshipit-source-id: 26d3ae58017fdcc70947ef29083903731e5c8d9b