Enable heapdom to write out the dominator tree just for a given set of types
Summary:
This follows up on the previous diff, which supported filtering objects directly dominated by the root node by a set of types, with a more comprehensive solution.

With this change, we support filtering the entire dominator tree by a set of types. This allows us to compute the aggregate size of (dominator subtrees of) objects for a given feature (say, belonging to a specific assembly or living in a specific namespace),

For example:
```
heapdom 'type "myassembly:" "myassembly.js"
```

Differential Revision: D54589241

fbshipit-source-id: f210d7925b100069f3c1046ac71fb1bb2e367f7b