Allow interactive object tagging
Summary:
When running a command such as
```
lo 'tagged 'destroyed 'exec "bt 'lifelines"
```
it can be very helpful to be able to quickly identify a specific set of objects in the resulting output (say, unowned object).

To this aim, we are introducing a new command `tag`, so that we can run something like:
```
lo 'unowned 'exec "tag 'unowned"
```
Now, the first command above will show the tag we added interactively.

Differential Revision: D54915135

fbshipit-source-id: 2d5c0ab7246813f6d13d27f585a94b8fd0913337