New flag bt 'counts to include number of instances per type in backtraces/lifelines
Summary: Provides a new option to include in backtraces the number of instances of the type mentioned on each line. This is particularly useful in lifelines, as it makes it easy to see when there is more than one instance of a type that is expected to be a singleton in a given scene.

Differential Revision: D55075511

fbshipit-source-id: e0273469b3ac9101d627cd7ab61300e0ca53dd66