Count total number of lifelines even when output limit reached
Summary: In order to be able to distinguish more easily in analysis results whether we have leaks piled upon leaks (and not just presence/absence of leaks), we can look at the number of lifelines found for all the leak candidates. The fact that this number was capped by the `'count` argument obscured these results, though, so we now report both number of lifelines found and number of lifelines printed.

Differential Revision: D56218518

fbshipit-source-id: 7d1e875caa369e0cb1bd4fdf05f5149bb09663c7